refactored: deleted course sorting
