--- frontend/src/constants/skillsMap.ts
@@ -1,2 +1,0 @@
-export const SET_FIRST = 1;
-export const SET_NEXT = -1;
