+++ frontend/src/pages/SkillsMap/StackMap/CourseMapForCurrentStack/CoursesList/index.tsx
@@ -1,1 +1,1 @@
-export { default } from './CoursesListContainer';
+export { default } from './CoursesList';
